backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
sentence-transformers==2.7.0
faiss-cpu==1.12.0
pypdf==4.3.1
google-generativeai==0.7.2
numpy==1.26.4
python-dotenv==1.0.0
reportlab==4.0.0
python-docx==0.8.12
sarvamai==0.1.7
speechrecognition==3.10.0
httpx==0.24.1

